Add peer connection through webrtc. Seperate components into different files
