+++ frontend/src/components/Header/Header.js
@@ -1,0 +1,9 @@
+const Header = () => {
+    return (
+        <div>
+            Nav Bar
+        </div>
+    );
+}
+
+export default Header;
